+++ src/app/shared/util.environment.ts
@@ -1,15 +1,11 @@
 import { createTokenList } from 'scripts/lib/createTokenList.mjs';
 import { environment } from 'src/environments/environment';
 
-export function createRegionSlugs(
-  regions: string = environment.regions,
-): string[] {
+export function createRegionSlugs(regions: string): string[] {
   return createTokenList(regions).map((s) => s.toLowerCase());
 }
 
-function createRegionLabels(
-  regionLabels: string = environment.regionsLabels,
-): string[] {
+function createRegionLabels(regionLabels: string): string[] {
   return createTokenList(regionLabels);
 }
 
@@ -29,8 +25,8 @@ export function getSheetIds(): { [key: string]: string } {
 }
 
 export function getRegionLabel(region: string = ''): string {
-  const regions = createRegionSlugs();
-  const regionsLabels = createRegionLabels();
+  const regions = createRegionSlugs(environment.regions);
+  const regionsLabels = createRegionLabels(environment.regionsLabels);
 
   let label = regionsLabels[regions.indexOf(region)];